feat(Placement): extract placement tab
Closes DCOS-18992
